Fin 50 handle errors

+++ back_end/controllers/category.js
@@ -2,7 +2,7 @@ import { Category } from '../models/Category.js';
 import { Transaction } from '../models/Transaction.js';
 import { Op } from 'sequelize';
 
-const listCategories = async (req, res) => {
+const listCategories = async (req, res, next) => {
   const { userId } = req.query;
 
   if (!userId) {
@@ -14,13 +14,13 @@ const listCategories = async (req, res) => {
         user_id: userId,
       },
     });
-    res.send(allCategories);
+    return res.send(allCategories);
   } catch (error) {
-    res.status(500).send(error);
+    next(error);
   }
 };
 
-const getCategory = async (req, res) => {
+const getCategory = async (req, res, next) => {
   const { id } = req.params;
   try {
     const category = await Category.findOne({
@@ -33,9 +33,9 @@ const getCategory = async (req, res) => {
       return res.status(404).json({ message: "Category doesn't exists" });
     }
 
-    res.send(category);
+    return res.send(category);
   } catch (error) {
-    res.status(500).send(error);
+    next(error);
   }
 };
 
@@ -91,13 +91,13 @@ const createCategory = async (req, res) => {
       color,
       user_id: userId,
     });
-    res.send(newCategory);
+    return res.send(newCategory);
   } catch (error) {
-    res.status(500).send(error);
+    next(error);
   }
 };
 
-const updateCategory = async (req, res) => {
+const updateCategory = async (req, res, next) => {
   const { id } = req.params;
   const { name, description, icon, color } = req.body;
 
@@ -118,23 +118,29 @@ const updateCategory = async (req, res) => {
 
     await category.save();
 
-    res.send(category);
+    return res.send(category);
   } catch (error) {
-    res.status(500).send(error);
+    next(error);
   }
 };
 
-const deleteCategory = async (req, res) => {
+const deleteCategory = async (req, res, next) => {
   const { id } = req.params;
+
   try {
-    await Category.destroy({
-      where: {
-        id: id,
-      },
-    });
-    res.send(204);
+    const category = await Category.findByPk(id);
+    if (!category) {
+      return res.status(400).send(`Category do not exists`);
+    } else {
+      const result = await Category.destroy({
+        where: {
+          id: id,
+        },
+      });
+      return res.status(200).send(`Category deleted ${result}`);
+    }
   } catch (error) {
-    res.status(500).send(error);
+    next(error);
   }
 };
